Keep deferred tool assertions local to compact tests
