+++ codex-rs/core/tests/suite/search_tool.rs
@@ -22,7 +22,6 @@ use core_test_support::apps_test_server::AppsTestServer;
 use core_test_support::apps_test_server::CALENDAR_CREATE_EVENT_MCP_APP_RESOURCE_URI;
 use core_test_support::apps_test_server::CALENDAR_CREATE_EVENT_RESOURCE_URI;
 use core_test_support::responses::ResponsesRequest;
-use core_test_support::responses::assert_tools_payload_does_not_defer;
 use core_test_support::responses::ev_assistant_message;
 use core_test_support::responses::ev_completed;
 use core_test_support::responses::ev_response_created;
@@ -607,7 +606,6 @@ async fn tool_search_returns_deferred_tools_without_follow_up_tool_injection() -
     );
 
     let first_request_body = requests[0].body_json();
-    assert_tools_payload_does_not_defer(&first_request_body);
     let first_request_tools = tool_names(&first_request_body);
     assert!(
         first_request_tools
@@ -827,7 +825,6 @@ async fn tool_search_returns_deferred_dynamic_tool_and_routes_follow_up_call() -
     assert_eq!(requests.len(), 3);
 
     let first_request_body = requests[0].body_json();
-    assert_tools_payload_does_not_defer(&first_request_body);
     let first_request_tools = tool_names(&first_request_body);
     assert!(
         first_request_tools
@@ -859,7 +856,6 @@ async fn tool_search_returns_deferred_dynamic_tool_and_routes_follow_up_call() -
     );
 
     let second_request_body = requests[1].body_json();
-    assert_tools_payload_does_not_defer(&second_request_body);
     let second_request_tools = tool_names(&second_request_body);
     assert!(
         !second_request_tools.iter().any(|name| name == tool_name),
@@ -878,7 +874,6 @@ async fn tool_search_returns_deferred_dynamic_tool_and_routes_follow_up_call() -
     );
 
     let third_request_body = requests[2].body_json();
-    assert_tools_payload_does_not_defer(&third_request_body);
     let third_request_tools = tool_names(&third_request_body);
     assert!(
         !third_request_tools.iter().any(|name| name == tool_name),
